+++ src/drv_mma845x.c
@@ -56,7 +56,7 @@ static uint8_t device_id;
 static sensor_align_e accAlign = CW90_DEG;
 
 static void mma8452Init(sensor_align_e align);
-static void mma8452Read(int16_t *accelData);
+static void mma8452Read(int32_t *accelData);
 
 bool mma8452Detect(sensor_t *acc)
 {
@@ -103,10 +103,10 @@ static void mma8452Init(sensor_align_e align)
         accAlign = align;
 }
 
-static void mma8452Read(int16_t *accelData)
+static void mma8452Read(int32_t *accelData)
 {
     uint8_t buf[6];
-    int16_t data[3];
+    int32_t data[3];
 
     i2cRead(MMA8452_ADDRESS, MMA8452_OUT_X_MSB, 6, buf);
     data[0] = ((int16_t)((buf[0] << 8) | buf[1]) >> 2) / 4;
