+++ src/drv_mpu6050.c
@@ -143,9 +143,9 @@ static sensor_align_e gyroAlign = CW0_DEG;
 static sensor_align_e accAlign = CW0_DEG;
 
 static void mpu6050AccInit(sensor_align_e align);
-static void mpu6050AccRead(int16_t *accData);
+static void mpu6050AccRead(int32_t *accData);
 static void mpu6050GyroInit(sensor_align_e align);
-static void mpu6050GyroRead(int16_t *gyroData);
+static void mpu6050GyroRead(int32_t *gyroData);
 
 extern uint16_t acc_1G;
 static uint8_t mpuAccelHalf = 0;
@@ -233,10 +233,10 @@ static void mpu6050AccInit(sensor_align_e align)
         accAlign = align;
 }
 
-static void mpu6050AccRead(int16_t *accData)
+static void mpu6050AccRead(int32_t *accData)
 {
     uint8_t buf[6];
-    int16_t data[3];
+    int32_t data[3];
 
     i2cRead(MPU6050_ADDRESS, MPU_RA_ACCEL_XOUT_H, 6, buf);
     data[0] = (int16_t)((buf[0] << 8) | buf[1]);
@@ -274,15 +274,19 @@ static void mpu6050GyroInit(sensor_align_e align)
         gyroAlign = align;
 }
 
-static void mpu6050GyroRead(int16_t *gyroData)
+static void mpu6050GyroRead(int32_t *gyroData)
 {
     uint8_t buf[6];
-    int16_t data[3];
+    int32_t data[3];
 
     i2cRead(MPU6050_ADDRESS, MPU_RA_GYRO_XOUT_H, 6, buf);
+
+    /* divide by 4 to give +-8192 == +-2000 degrees/sec */
+    
     data[0] = (int16_t)((buf[0] << 8) | buf[1]) / 4;
     data[1] = (int16_t)((buf[2] << 8) | buf[3]) / 4;
     data[2] = (int16_t)((buf[4] << 8) | buf[5]) / 4;
 
     alignSensors(data, gyroData, gyroAlign);
 }
+
